+++ src/seq_classes_ref.h
@@ -44,25 +44,25 @@ struct RefSequence {
         : name(""), nucleos(nucleos_) {};
 
     // Overloaded operator so nucleotides can be easily extracted
-    char operator[](const uint32& idx) const {
+    char operator[](const uint64& idx) const {
         if (idx >= nucleos.size()) {
             stop("Trying to extract nucleotide that doesn't exist");
         }
         return nucleos[idx];
     }
-    char& operator[](const uint32& idx) {
+    char& operator[](const uint64& idx) {
         if (idx >= nucleos.size()) {
             stop("Trying to extract nucleotide that doesn't exist");
         }
         return nucleos[idx];
     }
     // To resize this sequence
-    void reserve(const uint32& n) {
+    void reserve(const uint64& n) {
         nucleos.reserve(n);
         return;
     }
     // To resize this sequence
-    void resize(const uint32& n, const char& x) {
+    void resize(const uint64& n, const char& x) {
         // nucleos.resize(n, x); // the below way should be faster based on testing
         nucleos = std::string(n, x);
         return;
@@ -78,7 +78,7 @@ struct RefSequence {
         return;
     }
     // To return the size of this sequence
-    uint32 size() const noexcept {
+    uint64 size() const noexcept {
         return nucleos.size();
     }
     // For sorting from largest to smallest sequence
@@ -94,16 +94,16 @@ struct RefSequence {
      ------------------
      */
     void fill_read(std::string& read,
-                   const uint32& read_start,
-                   const uint32& seq_start,
-                   uint32 n_to_add) const {
+                   const uint64& read_start,
+                   const uint64& seq_start,
+                   uint64 n_to_add) const {
         // Making sure end doesn't go beyond the sequence bounds
         if ((seq_start + n_to_add - 1) >= nucleos.size()) {
             n_to_add = nucleos.size() - seq_start;
         }
         // Make sure the read is long enough (this fxn should never shorten it):
         if (read.size() < n_to_add + read_start) read.resize(n_to_add + read_start, 'N');
-        for (uint32 i = 0; i < n_to_add; i++) {
+        for (uint64 i = 0; i < n_to_add; i++) {
             read[(read_start + i)] = this->nucleos[(seq_start + i)];
         }
         return;
@@ -137,12 +137,12 @@ struct RefGenome {
     RefGenome(const RefGenome& ref_)
         : total_size(ref_.total_size), sequences(ref_.sequences),
           merged(ref_.merged), old_names(ref_.old_names) {};
-    RefGenome(const uint32& N)
+    RefGenome(const uint64& N)
         : sequences(std::deque<RefSequence>(N, RefSequence())) {};
     RefGenome(const std::deque<std::string>& seqs) {
-        uint32 n_seqs = seqs.size();
+        uint64 n_seqs = seqs.size();
         sequences = std::deque<RefSequence>(n_seqs, RefSequence());
-        for (uint32 i = 0; i < n_seqs; i++) {
+        for (uint64 i = 0; i < n_seqs; i++) {
             sequences[i].nucleos = seqs[i];
             sequences[i].name = "seq" + std::to_string(i);
             total_size += seqs[i].size();
@@ -150,26 +150,26 @@ struct RefGenome {
     }
     // Overloaded operator so sequences can be easily extracted
     // It returns a reference so no copying is done and so changes can be made
-    RefSequence& operator[](const uint32& idx) {
+    RefSequence& operator[](const uint64& idx) {
         if (idx >= sequences.size()) {
             stop("Trying to extract sequence that doesn't exist");
         }
         return sequences[idx];
     }
-    const RefSequence& operator[](const uint32& idx) const {
+    const RefSequence& operator[](const uint64& idx) const {
         if (idx >= sequences.size()) {
             stop("Trying to extract sequence that doesn't exist");
         }
         return sequences[idx];
     }
     // To return the number of sequences
-    uint32 size() const noexcept {
+    uint64 size() const noexcept {
         return sequences.size();
     }
     // To return the sequence sizes
-    std::vector<uint32> seq_sizes() const {
-        std::vector<uint32> out(size());
-        for (uint32 i = 0; i < out.size(); i++) out[i] = sequences[i].size();
+    std::vector<uint64> seq_sizes() const {
+        std::vector<uint64> out(size());
+        for (uint64 i = 0; i < out.size(); i++) out[i] = sequences[i].size();
         return out;
     }
     // For printing reference genome info
